requirements.txt
pytorch>=1.6.0
numpy
transformers>=4.2.2
networkx
tqdm
seaborn

